# LinearBVH optimizations