# Summary

- Optimizations for the `propagate_aabbs()` function:
  - Keep the BoundingBox for each thread in a local variable, so for each inner node we only need to access global memory for one of the child bounding boxes.
  - Use atomic operations to load and store the inner tree BoundingBoxes. Since atomics bypass the L1 cache and resolve at L2, this should avoid the inter-thread synchronization issues encountered in #307, without needing a call to `__threadfence()`.
- Optimizations for `bvh_traverse()`:
  - Based on [this](https://research.nvidia.com/publication/understanding-efficiency-ray-traversal-gpus-kepler-and-fermi-addendum), we use a "postpone slot" to store candidate leaves while allowing the traversal to continue. This seems to improve performance slightly by reducing warp divergence when performing the leaf operation.

Performance on 'mesh_tester', with car.stl:

| Phase | `develop` | this branch | factor improvement
| --------| ----------| ------------ | --------------------
| `propagate_aabbs()` | 80.349 ms | 1.610 ms | 49.9x
| `Fill[1]::count_traversal` | 9.217 ms | 9.855 ms | 0.94x
| `Fill[2]::fill_traversal` | 19.460 ms | 13.689 ms | 1.42x
| **Overall BVH::build** | 92.652 ms |  12.845 ms | 7.21x
| **Overall BVH::findBoundingBoxes** | 28.768 ms | 23.633 ms | 1.22x