@kanye-quest 2021-06-28
So after talking with Max Katz, it seems that there is, in fact, a possibility that the atomic write from the `atomicAdd` "store" operation is resolved after the `atomicAdd` from the "load" operation.

To work around this issue:
- The bounding boxes are initialized to their default, invalid states (`min_coord = FLT_MAX`, `max_coord = FLT_MIN`)
- On the GPU, the bounding box write uses `atomicExchange` to ensure it gets written out to L2. To read the bounding box written by another thread, for each float coordinate, we cast it to a `volatile` reference. Since standard loads of primitive types up to 128 bits shouldn't tear, we can poll the location for a value that isn't `FLT_MIN/MAX`.
- On the CPU, we can use `std::atomic_thread_fence` to ensure load/store order consistency.

I also added some #ifdef'ed debug code to track when polling was required to read a valid bounding box. It does seem that the `atomicAdd` to kill the first thread to reach an internal node seems to provide enough synchronization for the bounding box read/write in practice. But the polling check doesn't really cost much performance in the "happy" case, so might as well leave it in.